requirements.txt
mysql-connector-python==8.0.27
protobuf==3.19.1
pymongo==3.12.1
yapf==0.31.0